+++ controllers/front/notify.php
@@ -25,6 +25,8 @@ use Invertus\SaferPay\Api\Enum\TransactionStatus;
 use Invertus\SaferPay\Config\SaferPayConfig;
 use Invertus\SaferPay\Controller\AbstractSaferPayController;
 use Invertus\SaferPay\Core\Order\Action\UpdateOrderStatusAction;
+use Invertus\SaferPay\Core\Payment\DTO\CheckoutData;
+use Invertus\SaferPay\Core\SaferPayOrder\Action\UpdateSaferPayOrderAction;
 use Invertus\SaferPay\Processor\CheckoutProcessor;
 use Invertus\SaferPay\Repository\SaferPayOrderRepository;
 use Invertus\SaferPay\Service\SaferPayOrderStatusService;
@@ -37,6 +39,7 @@ if (!defined('_PS_VERSION_')) {
 class SaferPayOfficialNotifyModuleFrontController extends AbstractSaferPayController
 {
     const FILENAME = 'notify';
+    const SAFERPAY_ORDER_AUTHORIZE_ACTION = 'AUTHORIZE';
 
     /**
      * This code is being called by SaferPay by using NotifyUrl in InitializeRequest.
@@ -62,11 +65,15 @@ class SaferPayOfficialNotifyModuleFrontController extends AbstractSaferPayContro
         }
 
         try {
-            /** @var SaferPayTransactionAssertion $transactionAssert */
-            $transactionAssert = $this->module->getService(SaferPayTransactionAssertion::class);
-            $assertResponseBody = $transactionAssert->assert($cartId);
+            $assertResponseBody = $this->assertTransaction($cartId);
 
-            $this->authorizeSaferPayOrderEntity($cartId);
+            /** @var SaferPayOrderRepository $saferPayOrderRepository */
+            $saferPayOrderRepository = $this->module->getService(SaferPayOrderRepository::class);
+            $saferPayOrderId = $saferPayOrderRepository->getIdByCartId($cartId);
+
+            /** @var UpdateSaferPayOrderAction $updateSaferPayOrderAction */
+            $updateSaferPayOrderAction = $this->module->getService(UpdateSaferPayOrderAction::class);
+            $updateSaferPayOrderAction->run(new SaferPayOrder($saferPayOrderId), self::SAFERPAY_ORDER_AUTHORIZE_ACTION);
 
             // If order does not exist but assertion is valid that means order authorized or captured.
             $orderId = Order::getIdByCartId($cartId);
@@ -74,11 +81,14 @@ class SaferPayOfficialNotifyModuleFrontController extends AbstractSaferPayContro
             if (!$orderId) {
                 /** @var CheckoutProcessor $checkoutProcessor **/
                 $checkoutProcessor = $this->module->getService(CheckoutProcessor::class);
-                $checkoutProcessor
-                    ->processCreateOrderAfterAuthorization(
-                        $cart,
-                        $assertResponseBody->getPaymentMeans()->getBrand()->getPaymentMethod()
-                    );
+                $checkoutData = CheckoutData::createFromRequest(
+                    $cart,
+                    $assertResponseBody->getPaymentMeans()->getBrand()->getPaymentMethod(),
+                    Configuration::get(SaferPayConfig::IS_BUSINESS_LICENCE)
+                );
+                $checkoutData->setIsAuthorizedOrder(true);
+
+                $checkoutProcessor->run($checkoutData);
             }
 
             //TODO look into pipeline design pattern to use when object is modified in multiple places to avoid this issue.
@@ -87,7 +97,7 @@ class SaferPayOfficialNotifyModuleFrontController extends AbstractSaferPayContro
 
             /** @var UpdateOrderStatusAction $updateOrderStatusAction **/
             $updateOrderStatusAction = $this->module->getService(UpdateOrderStatusAction::class);
-            $updateOrderStatusAction->run($order, $assertResponseBody->getTransaction()->getStatus());
+            $updateOrderStatusAction->run((int) $orderId, (int) Configuration::get('SAFERPAY_PAYMENT_AUTHORIZED'));
 
             if (!$assertResponseBody->getLiability()->getLiabilityShift() &&
                 in_array($order->payment, SaferPayConfig::SUPPORTED_3DS_PAYMENT_METHODS) &&
@@ -129,25 +139,11 @@ class SaferPayOfficialNotifyModuleFrontController extends AbstractSaferPayContro
         die($this->module->l('Success', self::FILENAME));
     }
 
-    /**
-     * @param $cartId
-     * @return void
-     * @throws PrestaShopDatabaseException
-     * @throws PrestaShopException
-     */
-    private function authorizeSaferPayOrderEntity($cartId)
-    {
-        /** @var SaferPayOrderRepository $saferPayOrderRepository */
-        $saferPayOrderRepository = $this->module->getService(SaferPayOrderRepository::class);
-        $saferPayOrderId = $saferPayOrderRepository->getIdByCartId($cartId);
-        $saferPayOrder = new SaferPayOrder($saferPayOrderId);
+    private function assertTransaction($cartId) {
+        /** @var SaferPayTransactionAssertion $transactionAssert */
+        $transactionAssert = $this->module->getService(SaferPayTransactionAssertion::class);
 
-        if ($saferPayOrder->authorized) {
-            return;
-        }
-
-        $saferPayOrder->authorized = 1;
-        $saferPayOrder->update();
+        return $transactionAssert->assert(Order::getIdByCartId($cartId));
     }
 
     protected function displayMaintenancePage()